requirements.txt
alembic==1.5.8
appdirs==1.4.4
astroid==2.4.2
autopep8==1.5.4
bcrypt==3.2.0
black==20.8b1
cffi==1.14.4
click==7.1.2
coverage==5.5
cryptography==3.2.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-Migrate==2.7.0
Flask-MySQL==1.5.1
Flask-MySQLdb==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==2.0.1
pathspec==0.8.1
pep8==1.7.1
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
PyMySQL==0.10.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
regex==2020.11.13
six==1.15.0
SQLAlchemy==1.3.20
SQLAlchemy-Utils==0.36.8
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
wrapt==1.12.1